LTD: Fix wrong lock cost calculation
The lock price is currently calculated by using self.length, which
leds to a wrong behavior when the shop has more than 2 cards in the
Jokers area, if the user locks 2 and then unlocks all locked items,
and proceeds to purchase a card, the price will be set to 1 due to
current lock_list length being 0. This updates the calculation
logic, to only reduce price_mult by 1 if the card has been "locked".